connection to db

+++ src/main/java/ru/ispo/music_service/entity/PlaylistTrackId.java
@@ -6,6 +6,6 @@ import java.io.Serializable;
 
 @Embeddable
 class PlaylistTrackId implements Serializable {
-    private Long playlistId;
-    private Long trackId;
+    private Integer  playlistId;
+    private Integer  trackId;
 }
